+++ src/SW_netCDF_Output.c
@@ -376,6 +376,7 @@ static void write_pft_vals(int ncFileID, int varID, LOG_INFO *LogInfo) {
 @param[in] timeSize Number of time steps in current output slice
 @param[in] pd Current output netCDF period
 @param[in] startYr Start year of the simulation
+@param[in] posTimeInBnds Position of time coordinate values relative to bounds
 @param[out] bndsVals Start/end bounds for "time" variable; can be NULL
     if we are calculating the number of days from the base calendar year
 @param[out] dimVarVals Values of the "time" dimension; can be NULL
@@ -387,6 +388,7 @@ static void calc_num_timedays(
     size_t timeSize,
     OutPeriod pd,
     unsigned int startYr,
+    int posTimeInBnds,
     double *bndsVals,
     double *dimVarVals,
     double *startTime
@@ -437,9 +439,17 @@ static void calc_num_timedays(
         }
 
         if (!isnull(dimVarVals)) {
-            // time value = mid-time of bounds
-            dimVarVals[index] =
-                (bndsVals[index * 2] + bndsVals[index * 2 + 1]) / 2.0;
+            if (posTimeInBnds < 0) {
+                /* time value at start of bound */
+                dimVarVals[index] = bndsVals[index * 2];
+            } else if (posTimeInBnds == 0) {
+                /* time value at midpoint of bounds */
+                dimVarVals[index] =
+                    (bndsVals[index * 2] + bndsVals[index * 2 + 1]) / 2.0;
+            } else {
+                /* time value at end of bound */
+                dimVarVals[index] = bndsVals[index * 2 + 1];
+            }
         }
 
         *startTime += (double) numDays;
@@ -455,6 +465,7 @@ the variable "time_bnds" and fills the variable "time"
 @param[in] dimIDs Dimension identifiers for "vertical" and "bnds"
 @param[in] size Size of the vertical dimension/variable
 @param[in] dimVarID "time" dimension identifier
+@param[in] posTimeInBnds Position of time coordinate values relative to bounds
 @param[in] startYr Start year of the simulation
 @param[in,out] startTime Start number of days when dealing with
     years between netCDF files
@@ -468,6 +479,7 @@ static void create_time_vars(
     int dimIDs[],
     unsigned int size,
     int dimVarID,
+    int posTimeInBnds,
     unsigned int startYr,
     double *startTime,
     OutPeriod pd,
@@ -514,7 +526,13 @@ static void create_time_vars(
     }
 
     calc_num_timedays(
-        (size_t) size, pd, startYr, bndsVals, dimVarVals, startTime
+        (size_t) size,
+        pd,
+        startYr,
+        posTimeInBnds,
+        bndsVals,
+        dimVarVals,
+        startTime
     );
 
     SW_NC_write_vals(
@@ -547,6 +565,8 @@ the variable "vertical_bnds" and fills the variable "vertical"
 @param[in] hasConsistentSoilLayerDepths Flag indicating if all simulation
     run within domain have identical soil layer depths
     (though potentially variable number of soil layers)
+@param[in] posVerticalInBnds Position of vertical coordinate values
+    relative to bounds
 @param[in] lyrDepths Depths of soil layers (cm)
 @param[in] deflateLevel Level of deflation that will be used for the created
 variable
@@ -558,13 +578,14 @@ static void create_vert_vars(
     unsigned int size,
     int dimVarID,
     Bool hasConsistentSoilLayerDepths,
+    int posVerticalInBnds,
     const double lyrDepths[],
     int deflateLevel,
     LOG_INFO *LogInfo
 ) {
 
     double *dimVarVals = NULL;
-    double *bndVals = NULL;
+    double *bndsVals = NULL;
     double lyrStart = 0.0;
     const int numBnds = 2;
     size_t start[] = {0, 0};
@@ -593,7 +614,7 @@ static void create_vert_vars(
         return; // Exit function prematurely due to error
     }
 
-    bndVals = (double *) Mem_Malloc(
+    bndsVals = (double *) Mem_Malloc(
         (size_t) (size * numBnds) * sizeof(double), "create_vert_vars", LogInfo
     );
     if (LogInfo->stopRun) {
@@ -602,16 +623,28 @@ static void create_vert_vars(
     }
 
     for (size_t index = 0; index < (size_t) size; index++) {
+
+        bndsVals[index * 2] = lyrStart;
+
         // if hasConsistentSoilLayerDepths,
         // then use soil layer depth, else soil layer number
-        dimVarVals[index] = (hasConsistentSoilLayerDepths) ?
-                                lyrDepths[index] :
-                                (double) (index + 1);
+        bndsVals[index * 2 + 1] = (hasConsistentSoilLayerDepths) ?
+                                      lyrDepths[index] :
+                                      (double) (index + 1);
 
-        bndVals[index * 2] = lyrStart;
-        bndVals[index * 2 + 1] = dimVarVals[index];
+        lyrStart = bndsVals[index * 2 + 1];
 
-        lyrStart = bndVals[index * 2 + 1];
+        if (posVerticalInBnds < 0) {
+            /* vertical value at shallow/top of bound */
+            dimVarVals[index] = bndsVals[index * 2];
+        } else if (posVerticalInBnds == 0) {
+            /* vertical value at midpoint of bounds */
+            dimVarVals[index] =
+                (bndsVals[index * 2] + bndsVals[index * 2 + 1]) / 2.0;
+        } else {
+            /* vertical value at deep/bottom of bound */
+            dimVarVals[index] = bndsVals[index * 2 + 1];
+        }
     }
 
     SW_NC_write_vals(
@@ -619,7 +652,7 @@ static void create_vert_vars(
     );
     free(dimVarVals);
     if (LogInfo->stopRun) {
-        free(bndVals);
+        free(bndsVals);
         return; // Exit function prematurely due to error
     }
 
@@ -629,14 +662,14 @@ static void create_vert_vars(
         &bndIndex,
         ncFileID,
         "vertical_bnds",
-        bndVals,
+        bndsVals,
         start,
         count,
         "double",
         LogInfo
     );
 
-    free(bndVals);
+    free(bndsVals);
 }
 
 /**
@@ -655,9 +688,12 @@ if needed
 @param[in] hasConsistentSoilLayerDepths Flag indicating if all simulation
     run within domain have identical soil layer depths
     (though potentially variable number of soil layers)
+@param[in] posVerticalInBnds Position of vertical coordinate values
+    relative to bounds
 @param[in] lyrDepths Depths of soil layers (cm)
 @param[in,out] startTime Start number of days when dealing with
     years between netCDF files
+@param[in] posTimeInBnds Position of time coordinate values relative to bounds
 @param[in] startYr Start year of the simulation
 @param[in] pd Current output netCDF period
 @param[in] deflateLevel Level of deflation that will be used for the created
@@ -670,9 +706,11 @@ static void fill_dimVar(
     unsigned int size,
     int varID,
     Bool hasConsistentSoilLayerDepths,
+    int posVerticalInBnds,
     double lyrDepths[],
     double *startTime,
     int dimNum,
+    int posTimeInBnds,
     unsigned int startYr,
     OutPeriod pd,
     int deflateLevel,
@@ -707,6 +745,7 @@ static void fill_dimVar(
                     size,
                     varID,
                     hasConsistentSoilLayerDepths,
+                    posVerticalInBnds,
                     lyrDepths,
                     deflateLevel,
                     LogInfo
@@ -719,6 +758,7 @@ static void fill_dimVar(
                     dimIDs,
                     size,
                     varID,
+                    posTimeInBnds,
                     startYr,
                     startTime,
                     pd,
@@ -1272,7 +1312,9 @@ static void create_output_file(
                 (const char **) attVals,
                 numAtts,
                 hasConsistentSoilLayerDepths,
+                OutDom->netCDFOutput.posVerticalInBnds,
                 lyrDepths,
+                OutDom->netCDFOutput.posTimeInBnds,
                 startTime,
                 baseCalendarYear,
                 startYr,
@@ -1331,7 +1373,10 @@ and fill the variable with the respective information
 @param[in] hasConsistentSoilLayerDepths Flag indicating if all simulation
     run within domain have identical soil layer depths
     (though potentially variable number of soil layers)
+@param[in] posVerticalInBnds Position of vertical coordinate values
+    relative to bounds
 @param[in] lyrDepths Depths of soil layers (cm)
+@param[in] posTimeInBnds Position of time coordinate values relative to bounds
 @param[in,out] startTime Start number of days when dealing with
     years between netCDF files (returns updated value)
 @param[in] baseCalendarYear First year of the entire simulation
@@ -1347,7 +1392,9 @@ void SW_NCOUT_create_output_dimVar(
     int ncFileID,
     int *dimID,
     Bool hasConsistentSoilLayerDepths,
+    int posVerticalInBnds,
     double lyrDepths[],
+    int posTimeInBnds,
     double *startTime,
     unsigned int baseCalendarYear,
     unsigned int startYr,
@@ -1448,9 +1495,11 @@ void SW_NCOUT_create_output_dimVar(
             size,
             varID,
             hasConsistentSoilLayerDepths,
+            posVerticalInBnds,
             lyrDepths,
             startFillTime,
             dimNum,
+            posTimeInBnds,
             startYr,
             pd,
             deflateLevel,
@@ -1955,6 +2004,9 @@ void SW_NCOUT_init_ptrs(SW_NETCDF_OUT *SW_netCDFOut) {
     SW_netCDFOut->strideOutYears = -1;
     SW_netCDFOut->deflateLevel = 0;
 
+    SW_netCDFOut->posTimeInBnds = 0;     /* default: centered */
+    SW_netCDFOut->posVerticalInBnds = 1; /* default: bottom bound */
+
     for (index = 0; index < numAllocVars; index++) {
         *allocArr[index] = NULL;
     }
@@ -2336,8 +2388,9 @@ void SW_NCOUT_create_output_files(
                 timeSize,
                 pd,
                 (IntU) baseCalendarYear,
-                NULL,
-                NULL,
+                0,    // unused
+                NULL, // unused
+                NULL, // unused
                 &baseStartTime[pd]
             );
         }
@@ -3092,14 +3145,16 @@ void SW_NCOUT_read_atts(
         "geo_YAxisName",
         "proj_XAxisName",
         "proj_YAxisName",
-        "siteName"
+        "siteName",
+        "posTimeInBnds",
+        "posVerticalInBnds"
     };
     static const Bool requiredKeys[NUM_ATT_IN_KEYS] = {
-        swTRUE,  swTRUE,  swTRUE,  swFALSE, swFALSE, swTRUE,  swTRUE,
-        swTRUE,  swTRUE,  swTRUE,  swTRUE,  swTRUE,  swTRUE,  swTRUE,
-        swFALSE, swFALSE, swFALSE, swFALSE, swFALSE, swFALSE, swFALSE,
-        swFALSE, swFALSE, swFALSE, swFALSE, swFALSE, swFALSE, swFALSE,
-        swTRUE,  swTRUE,  swTRUE,  swTRUE,  swTRUE,  swTRUE
+        swTRUE,  swTRUE,  swTRUE,  swFALSE, swFALSE, swTRUE,  swTRUE,  swTRUE,
+        swTRUE,  swTRUE,  swTRUE,  swTRUE,  swTRUE,  swTRUE,  swFALSE, swFALSE,
+        swFALSE, swFALSE, swFALSE, swFALSE, swFALSE, swFALSE, swFALSE, swFALSE,
+        swFALSE, swFALSE, swFALSE, swFALSE, swTRUE,  swTRUE,  swTRUE,  swTRUE,
+        swTRUE,  swTRUE,  swFALSE, swFALSE
     };
     Bool hasKeys[NUM_ATT_IN_KEYS] = {swFALSE};
 
@@ -3173,7 +3228,8 @@ void SW_NCOUT_read_atts(
         // set_hasKey() does not produce errors, only warnings possible
 
         /* Check to see if the line number contains a double or integer value */
-        doIntConv = (Bool) (keyID >= 25 && keyID <= 29);
+        doIntConv =
+            (Bool) (keyID >= 25 && keyID <= 29) || (keyID >= 35 && keyID <= 36);
         doDoubleConv = (Bool) ((keyID >= 10 && keyID <= 12) ||
                                (keyID >= 17 && keyID <= 19) ||
                                (keyID >= 23 && keyID <= 24));
@@ -3363,6 +3419,12 @@ void SW_NCOUT_read_atts(
             break;
         case 34:
             SW_netCDFOut->siteName = Str_Dup(value, LogInfo);
+            break;
+        case 35:
+            SW_netCDFOut->posTimeInBnds = inBufintRes;
+            break;
+        case 36:
+            SW_netCDFOut->posVerticalInBnds = inBufintRes;
             break;
         case KEY_NOT_FOUND:
         default: